requirements.txt
openai==0.25.0
Pillow==9.3.0
streamlit==1.16.0
